修改配置file-service文件

+++ filebox-common/src/main/java/top/xufilebox/common/util/Constant.java
@@ -6,7 +6,7 @@ package top.xufilebox.common.util;
  * @create: 2020-12-20 20:23
  **/
 public class Constant {
-    public static final int DEFAULT_CAPACITY = 1073741824;
+    public static final Long DEFAULT_CAPACITY = 1073741824L;
     public static final int DEFAULT_PARENT_DIR_ID = -1;
     public static final int UPLOAD_COMPLETED = 1;
     public static final int UPLOAD_UNCOMPLETED = 0;
